objects/Door: Fix GenericType(_New)
Must be new field in 1.69

+++ lib/nw/objects/Door.hpp
@@ -62,7 +62,7 @@ struct Door : public ObjectBase {
     DoorAnimationState animation_state = DoorAnimationState::closed;
     std::string linked_to;
     uint16_t loadscreen = 0;
-    uint8_t generic_type = 0;
+    uint32_t generic_type = 0;
     uint8_t linked_to_flags = 0;
 
     uint32_t appearance;
